@@ -8,15 +8,19 @@
   ],
   "license": "MIT",
   "require": {
-    "php": ">=7.3",
-    "plaisio/babel-core": "^1.3",
+    "php": ">=8.1",
+    "plaisio/babel-core": "^2.0.0",
     "plaisio/kernel": "^3.0",
-    "plaisio/language-resolver": "^1.2"
+    "plaisio/language-resolver": "^1.2",
+    "plaisio/request": "^0.13.1"
   },
+  "minimum-stability": "dev",
+  "prefer-stable": true,
   "require-dev": {
-    "phing/phing": "^2.0",
-    "phpunit/phpunit": "^9.0",
-    "plaisio/console": "^2.0"
+    "phing/phing": "^3.0.0-RC4",
+    "phpunit/phpunit": "^9.6.3",
+    "plaisio/console-kernel": "^1.0.1",
+    "plaisio/request-core": "^0.13.0"
   },
   "autoload": {
     "psr-4": {
@@ -30,6 +34,9 @@
   },
   "config": {
     "bin-dir": "bin/",
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "phing/phing-composer-configurator": true
+    }
   }
 }
