@@ -19,7 +19,7 @@ class CoreLanguageResolverTest extends TestCase
    *
    * @var PlaisioKernel
    */
-  private $kernel;
+  private PlaisioKernel $kernel;
 
   //--------------------------------------------------------------------------------------------------------------------
   /**
